+++ test/src/interface/subscription_repository_test.dart
@@ -3,7 +3,6 @@ import 'dart:convert';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:twitch_client/src/interface/subscription_repository.dart';
-import 'package:twitch_client/src/response/get_broadcaster_subscription_response.dart';
 import 'package:twitch_client/twitch_client.dart';
 
 import '../../json/asset_reader.dart';